// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AuthMetod = {
  CREDENTIALS: 'CREDENTIALS',
  GOOGLE: 'GOOGLE',
  GITHUB: 'GITHUB'
} as const

export type AuthMetod = (typeof AuthMetod)[keyof typeof AuthMetod]


export const TokenType = {
  VERIFICATION: 'VERIFICATION',
  TWO_FACTOR: 'TWO_FACTOR',
  PASSWORD_RESET: 'PASSWORD_RESET'
} as const

export type TokenType = (typeof TokenType)[keyof typeof TokenType]


export const Theme = {
  LIGHT: 'LIGHT',
  DARK: 'DARK'
} as const

export type Theme = (typeof Theme)[keyof typeof Theme]


export const Language = {
  ENGLISH: 'ENGLISH',
  RUSSIAN: 'RUSSIAN'
} as const

export type Language = (typeof Language)[keyof typeof Language]